Fixed a bug if the output file was the input file and was a single image MRC file (the output file had been deleted to fix another problem).

+++ programs/e2proc2d.py
@@ -268,7 +268,10 @@ def main():
 
 					if out_ext == ".mrc" :
 						if os.path.isfile(outfile) :
-							os.remove(outfile)
+							if infile == outfile :
+								options.inplace = True
+							else :
+								os.remove(outfile)
 		else :
 			num_inp_images = -1
 			print "Input file '" + infile + "' does not exist."
